added affix counter and better logic for alt auging(default)
Added affix counter bassed off what affix information we are giving it
to begin with.
Based off that, we can now know if its worth to aug or not.

E.G (default)
want a suffix but we roll 1 suffix and it isnt what we want, alt it
want a suffix and we roll 1 prefix and no suffix, aug it

maybe other stuff i cant remember